refactor(typescript): move browserInfoHelper tests to typescript
